+++ src/rtc/dynamic_h264_video_encoder.cpp
@@ -82,7 +82,7 @@ VideoFrameType ConvertToVideoFrameType(EVideoFrameType type) {
     case videoFrameTypeInvalid:
       break;
   }
-  RTC_NOTREACHED() << "Unexpected/invalid frame type: " << type;
+  RTC_DCHECK_NOTREACHED() << "Unexpected/invalid frame type: " << type;
   return VideoFrameType::kEmptyFrame;
 }
 
@@ -200,8 +200,9 @@ void DynamicH264VideoEncoder::ReleaseOpenH264() {
   }
 }
 
-int32_t DynamicH264VideoEncoder::InitEncode(const VideoCodec* inst,
-                                    const VideoEncoder::Settings& settings) {
+int32_t DynamicH264VideoEncoder::InitEncode(
+    const VideoCodec* inst,
+    const VideoEncoder::Settings& settings) {
   ReportInit();
   if (!inst || inst->codecType != kVideoCodecH264) {
     ReportError();
@@ -355,7 +356,8 @@ int32_t DynamicH264VideoEncoder::RegisterEncodeCompleteCallback(
   return WEBRTC_VIDEO_CODEC_OK;
 }
 
-void DynamicH264VideoEncoder::SetRates(const RateControlParameters& parameters) {
+void DynamicH264VideoEncoder::SetRates(
+    const RateControlParameters& parameters) {
   if (encoders_.empty()) {
     RTC_LOG(LS_WARNING) << "SetRates() while uninitialized.";
     return;
@@ -524,7 +526,8 @@ int32_t DynamicH264VideoEncoder::Encode(
     if (encoded_images_[i].size() > 0) {
       // Parse QP.
       h264_bitstream_parser_.ParseBitstream(encoded_images_[i]);
-      encoded_images_[i].qp_ = h264_bitstream_parser_.GetLastSliceQp().value_or(-1);
+      encoded_images_[i].qp_ =
+          h264_bitstream_parser_.GetLastSliceQp().value_or(-1);
 
       // Deliver encoded image.
       CodecSpecificInfo codec_specific;
@@ -567,7 +570,7 @@ SEncParamExt DynamicH264VideoEncoder::CreateEncoderParams(size_t i) const {
   } else if (codec_.mode == VideoCodecMode::kScreensharing) {
     encoder_params.iUsageType = SCREEN_CONTENT_REAL_TIME;
   } else {
-    RTC_NOTREACHED();
+    RTC_DCHECK_NOTREACHED();
   }
   encoder_params.iPicWidth = configurations_[i].width;
   encoder_params.iPicHeight = configurations_[i].height;
@@ -610,8 +613,8 @@ SEncParamExt DynamicH264VideoEncoder::CreateEncoderParams(size_t i) const {
   if (encoder_params.iTemporalLayerNum > 1) {
     encoder_params.iNumRefFrame = 1;
   }
-  RTC_LOG(INFO) << "OpenH264 version is " << OPENH264_MAJOR << "."
-                << OPENH264_MINOR;
+  RTC_LOG(LS_INFO) << "OpenH264 version is " << OPENH264_MAJOR << "."
+                   << OPENH264_MINOR;
   switch (packetization_mode_) {
     case H264PacketizationMode::SingleNalUnit:
       // Limit the size of the packets produced.
@@ -620,8 +623,8 @@ SEncParamExt DynamicH264VideoEncoder::CreateEncoderParams(size_t i) const {
           SM_SIZELIMITED_SLICE;
       encoder_params.sSpatialLayers[0].sSliceArgument.uiSliceSizeConstraint =
           static_cast<unsigned int>(max_payload_size_);
-      RTC_LOG(INFO) << "Encoder is configured with NALU constraint: "
-                    << max_payload_size_ << " bytes";
+      RTC_LOG(LS_INFO) << "Encoder is configured with NALU constraint: "
+                       << max_payload_size_ << " bytes";
       break;
     case H264PacketizationMode::NonInterleaved:
       // When uiSliceMode = SM_FIXEDSLCNUM_SLICE, uiSliceNum = 0 means auto
@@ -659,7 +662,6 @@ VideoEncoder::EncoderInfo DynamicH264VideoEncoder::GetEncoderInfo() const {
   info.scaling_settings =
       VideoEncoder::ScalingSettings(kLowH264QpThreshold, kHighH264QpThreshold);
   info.is_hardware_accelerated = false;
-  info.has_internal_source = false;
   info.supports_simulcast = true;
   // TODO(melpon): M88 あたりでコメントインする
   //info.preferred_pixel_formats = {VideoFrameBuffer::Type::kI420};