+++ src/zakuro.cpp
@@ -26,7 +26,7 @@
 
 Zakuro::Zakuro(ZakuroConfig config) : config_(std::move(config)) {}
 
-const int MESSAGE_SIZE_MIN = 22;
+const int MESSAGE_SIZE_MIN = 48;
 const int MESSAGE_SIZE_MAX = 256 * 1000;
 const int BINARY_POOL_SIZE = 1 * 1024 * 1024;
 
@@ -228,9 +228,10 @@ int Zakuro::Run() {
   rtcm_config.simulcast = config_.sora_simulcast;
   rtcm_config.priority = config_.priority;
   rtcm_config.openh264 = config_.openh264;
-  rtcm_config.use_dcsctp = true;
   rtcm_config.fake_network_send = config_.fake_network_send;
   rtcm_config.fake_network_receive = config_.fake_network_receive;
+  rtcm_config.initial_mute_video = config_.initial_mute_video;
+  rtcm_config.initial_mute_audio = config_.initial_mute_audio;
   if (config_.no_audio_device) {
     rtcm_config.audio_type = RTCManagerConfig::AudioType::NoAudio;
   } else if (!config_.game.empty()) {
@@ -287,6 +288,8 @@ int Zakuro::Run() {
 
     SoraClientConfig sorac_config;
     sorac_config.insecure = config_.insecure;
+    sorac_config.client_cert = config_.client_cert;
+    sorac_config.client_key = config_.client_key;
     sorac_config.signaling_urls = config_.sora_signaling_urls;
     sorac_config.channel_id = config_.sora_channel_id;
     sorac_config.video = config_.sora_video;